Primate_Joystick_Pull/requirements.txt
Pillow==9.5.0
# pyscreenshot==3.0
hjson==3.0.2
